requirement.txt
mysql-connector-python==8.0.27
matplotlib==3.5.1
tk=3.10.1